This is the WeMos USB port driver if needed
I found this on the internet when my Windows 10 laptop stopped seeing the WeMos D1 as a valid com port. For some reason my Windows 10 machine said the USB connection but could not find the driver.

I inserted the USB cable into the PC and WeMos first, then unzipped and installed this driver.
Once installed the Arduino programming tool sees the WeMos on a valid com port.